opera_hotels_reports/requirements.txt
tabula-py
pandas
openpyxl
XlsxWriter